requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
langchain-community
python-dotenv
google-generativeai
faiss-cpu
transformers
sentence-transformers
deep_translator
scikit-learn
matplotlib
seaborn
gtts
torch
requests
passlib